Fix num_qubits() and num_clbits() methods (#122)
The num bits methods were incorrectly querying the target stored with
the circuit, when present, for the number of bits instead of the actual
circuit attributes. This commit fixes this issue so the circuit bit
counts are correctly returned.

There is a larger question around why the circuit is locally storing the
bit counts instead of just querying the inner qiskit circuit object. But
for just fixing this issue I kept the change minimal rather than
removing the protected attribute. We can do that in a follow up if
needed.