Fix bug issue template environment question (#126)
This commit fixes the bug issue template in the repo to be for this
repo. It appears the template was copied from:

https://github.com/Qiskit/direct-access-client

and not updated to reflect the information needed for Qiskit-C++. This
commit updates the environment questions to collect more relevant
details of the system environment to aid in triaging bug reports.